# Add unit tests for encoder MC