--- a/test/build/win32/codec_ut/codec_unittest.vcproj
+++ b/test/build/win32/codec_ut/codec_unittest.vcproj
@@ -415,6 +415,10 @@
 				>
 			</File>
 			<File
+				RelativePath="..\..\..\encoder\EncUT_MotionCompensation.cpp"
+				>
+			</File>
+			<File
 				RelativePath="..\..\..\encoder\EncUT_MotionEstimate.cpp"
 				>
 			</File>