The encoder contains separate (and separately tweaked) copies of the MC functions compared to the decoder. Add tests for these functions. Also simplify the structure of the MC tests a little bit (speeding up the compile time and reducing the overall number of tests).

Review at https://rbcommons.com/s/OpenH264/r/1090/.