@sijchen 2015-01-27
can the failed CI record be remove from this pull request?