@sijchen 2018-01-27
Archiving from https://rbcommons.com/s/OpenH264/r/1090/

----
**testing_done**:  
**description**:  <p>The encoder contains separate (and separately tweaked) copies of the MC functions compared to the decoder. Add tests for these functions. Also simplify the structure of the MC tests a little bit (speeding up the compile time and reducing the overall number of tests).</p><p></p><p>Simplify code in HorFilterInput16bit in MC</p><p>This avoids a gcc optimizer bug (which seems to be present in some<br />gcc 4.6 and 4.7 versions) at the -O3 level.</p><p></p><p>Add const to some inline functions within MC</p><p></p><p>Bundle MC tests for all motion vectors in one single test</p><p>This speeds up the compile time from 21.3 to 2.6 seconds<br />for the MC test files.</p><p>This makes it slightly harder to see exactly which test<br />failed on a quick glance, but it makes the overall structure of<br />the unit test output more manageable and readable, by reducing<br />the number of tests from 1300 to 430.</p><p>Add unit tests for special cased MC functions in the encoder</p><p></p><p>Add unit tests for encoder MC</p><p></p><p>Prepare the shared MC test code for adding an encoder MC test</p><p></p><p>Split DecUT_MotionCompensation to a shareable header</p><p>This allows adding a version of the same test for the encoder.</p><p>Make local functions and tables in DecUT_MotionCompensation static</p><p></p><p>Unify the pfChromaMc function signature with the decoder version</p><p>Instead of passing a struct with x/y, pass them as two separate<br />parameters.</p><p>Remove unused includes in DecUT_MotionCompensation</p><p></p><p>Unify the order of iX and iY in DecUT_MotionCompensation</p><p>They are in the order iX, iY in the rest of the file.</p><p>Avoid a duplicate local typedef</p><p></p><p>Unify the encoder MC init function with the decoder</p><p>This simplifies the code a little, by passing a pointer to the sub-struct<br />SMcFunc instead of to the full SWelsFuncPtrList, which isn't necessary.</p>

----
2015-01-27T00:09:38+03:00 status-change:
Added a workaround for an optimizer bug in gcc 4.6/4.7
The encoder contains separate (and separately tweaked) copies of the MC functions compared to the decoder. Add tests for these functions. Also simplify the structure of the MC tests a little bit (speeding up the compile time and reducing the overall number of tests).
Simplify code in HorFilterInput16bit in MC
This avoids a gcc optimizer bug (which seems to be present in some<br/></td></tr><tr class="insert"> <td class="marker">&nbsp;</td> <td class="marker">+</td> <td class="line rich-text">gcc 4.6 and 4.7 versions) at the -O3 level.
Add const to some inline functions within MC
Bundle MC tests for all motion vectors in one single test
This speeds up the compile time from 21.3 to 2.6 seconds<br/></td></tr><tr class="equal"> <td class="marker">&nbsp;</td> <td class="marker">&nbsp;</td> <td class="line rich-text">for the MC test files.
This makes it slightly harder to see exactly which test<br/></td></tr><tr class="equal"> <td class="marker">&nbsp;</td> <td class="marker">&nbsp;</td> <td class="line rich-text">failed on a quick glance, but it makes the overall structure of<br/></td></tr><tr class="equal"> <td class="marker">&nbsp;</td> <td class="marker">&nbsp;</td> <td class="line rich-text">the unit test output more manageable and readable, by reducing<br/></td></tr><tr class="equal"> <td class="marker">&nbsp;</td> <td class="marker">&nbsp;</td> <td class="line rich-text">the number of tests from 1300 to 430.
Add unit tests for special cased MC functions in the encoder
Add unit tests for encoder MC
Prepare the shared MC test code for adding an encoder MC test
Split DecUT_MotionCompensation to a shareable header
This allows adding a version of the same test for the encoder.
Make local functions and tables in DecUT_MotionCompensation static
Unify the pfChromaMc function signature with the decoder version
Instead of passing a struct with x/y, pass them as two separate<br/></td></tr><tr class="equal"> <td class="marker">&nbsp;</td> <td class="marker">&nbsp;</td> <td class="line rich-text">parameters.
Remove unused includes in DecUT_MotionCompensation
Unify the order of iX and iY in DecUT_MotionCompensation
They are in the order iX, iY in the rest of the file.
Avoid a duplicate local typedef
Unify the encoder MC init function with the decoder
This simplifies the code a little, by passing a pointer to the sub-struct<br/></td></tr><tr class="equal"> <td class="marker">&nbsp;</td> <td class="marker">&nbsp;</td> <td class="line rich-text">SMcFunc instead of to the full SWelsFuncPtrList, which isn't necessary.


----
 reviewer dong zhang (dongzhang) 
 time 2015-01-27T10:03:20+03:00 
2015-01-27T10:03:20+03:00 dong zhang (dongzhang):
>Ship It!
Ship It!
Ship it!


----

----

Finish archiving from https://rbcommons.com/s/OpenH264/r/1090/